refactor: Update Navbar component to include "Entrar" button
The code changes in this commit update the Navbar component to include an "Entrar" button. This button allows users to login to the application. Additionally, the "Mudar visualização" button is moved to the right side of the Navbar. This improves the user interface and enhances the navigation experience within the application.