refactor: Add "Favoritar" button to FilmCard component
This commit adds a "Favoritar" button to the FilmCard component. The button allows users to save a film as a favorite. When clicked, the onSaveFavorite function is called with the film data as a parameter. This enhancement improves the user experience by providing a convenient way to mark films as favorites.